Modify SyncBN2BN.
