+++ mmseg/ops/__init__.py
@@ -1,5 +1,4 @@
 from .encoding import Encoding
-from .sync2bn import revert_sync_batchnorm
 from .wrappers import Upsample, resize
 
-__all__ = ['Upsample', 'resize', 'Encoding', 'revert_sync_batchnorm']
+__all__ = ['Upsample', 'resize', 'Encoding']
